Don't hide values of builtins (issue8)
When a statement gets reinterpreted the current implementation goes
out of it's way to to not show the result of the call but rather the
call itself.  This means things like "len(l) == 42" will be shown as
"len([0, ...]) == 42" instead of the more desirable "666 == 42".  So
this patch reverts that behaviour to show the actual results of
builtins again.