Add tests for unicode and fix overzealous unicode casting
There is no reason to force ''.join([...]) to be a unicode string when
the items in the list could be unicode.  If an argument is unicode py2
does convert all of it to unicode implicitly anyway and the explict
casting does destroy smuggling of non-ascii bytes as
test_unicode_handling() does.