Only special treat assertion errors with a msg attribute
All other exceptions should not get a special treatment because we are
not in charge of their text.  The formatted exception should also include
the class name which we can just hardcode instead of tying to use the class
name since using the class name is more complicated as the builtin assertion
error gets replaced.