make_numbered_dir algorithm improved to make it multi-process safe by means of .lock file utilization.
There are two kinds of actors: user(single) and remover(many).
User, having created .lock in directory can be sure that he's the owner
of directory. Remover, having created .lock in directory, is sure that
he's the only owner of directory. Only safe operation for remover to do
is to rename to-be-removed, exlusively-owned directory, pattern of which
is used by remover actors alone. This is needed to avoid situation when
two removers schedule removal of certain directory, one is preempted
and incorrectly removes newly created directory of some working user.