Fix travis stages
As it were, "deploy" would run before "test" because it was left out of the
stages section.

Instead of running "test" and "deploy" stages on tags, to run just
"deploy"; we already require that all tests have passed before someone pushes
a tag for deployment anyway, no sense running the tests again.